Django_exercise1_1/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.10.0
djangorestframework==3.12.2
Pillow==8.0.1
Pygments==2.7.3
pytz==2020.4
sqlparse==0.4.1
